save capillaries ok; pb with export measures to excel
save capillaries: initially, modified capillaries were "copy"-ed; now,
the roi within a capillary is made = to roi. It seems to work while it
did not work before - but I don't know why the "copy" result is erratic.

excel export: pb with list of files which is empty now, so that it is
impossible to get the time at which a file was created